Projects/news_letter/requirements.txt
selenium==4.10.0
numpy==1.24.3
matplotlib==3.7.1
pandas==2.0.2
reportlab==4.0.4
PyPDF2==3.0.1
fpdf2==2.7.4